requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
SQLAlchemy>=2.0.23
alembic>=1.13.1
pydantic>=1.10.13,<2.0
python-dotenv>=1.0.1
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
redis>=5.0.1
loguru>=0.7.2
email-validator>=2.1.0
httpx>=0.25.2
pytest>=7.4.4
pytest-asyncio>=0.23.3
fakeredis>=2.21.3
